+++ lib/interactive_tree_view.dart
@@ -1,5 +1,6 @@
 library;
 
 export 'src/treeview.dart';
+export 'src/static_treeview.dart';
 export 'src/treecontroller.dart';
 export 'src/defaults.dart';