+++ CHANGELOG.md
@@ -8,6 +8,8 @@
  - Added new onChanged callback to controller.
  - Added replaceWith method to `TreeNode`
  - Removed the requirement for the attached node to be not attached.
+ - Added `StaticTreeView` widget
+ - made internal file structure change
 ## 0.9.2
  - Added the option to drag nodes as a child of an empty node, by hovering over the far right of it.
  - Added child option to placement enum